[IMP] - add scoring rules

+++ src/main/java/ai/qorva/core/dto/common/ScoringWeight.java
@@ -1,10 +1,8 @@
 package ai.qorva.core.dto.common;
 
-import java.math.BigDecimal;
-
 public record ScoringWeight(
-        BigDecimal skills,
-        BigDecimal experience,
-        BigDecimal location,
-        BigDecimal industry
+        Double skills,
+        Double experience,
+        Double location,
+        Double industry
 ) {}
