+++ src/main/java/ai/qorva/core/dto/common/SkillRequirement.java
@@ -1,11 +1,9 @@
 package ai.qorva.core.dto.common;
 
-import java.math.BigDecimal;
-
 public record SkillRequirement(
         String name,
         String importance,
-        BigDecimal weight,
+        Double weight,
         Integer minYearsOfExperience,
         Boolean exactSkillOnly
 ) {}
